+++ layouts/default.vue
@@ -12,9 +12,6 @@
 
 <script setup lang="ts">
 const open = ref(false)
-const store = useStore()
-await useAsyncData('jobs', () => store.fetchJobs())
-await useAsyncData('space', () => store.fetchSpace())
 </script>
 
 <style lang="scss">