+++ pages/experiences/index.vue
@@ -1,14 +1,14 @@
 <template>
   <ComingSoon v-if="!page">{{ $t('experiences.title') }}</ComingSoon>
   <template v-else>
-    <div class="container-page">
-      <header class="section-experiences__title">
-        <BaseHeading primary :label="$t('experiences.subtitle')">{{ $t('experiences.title') }}</BaseHeading>
-      </header>
-    </div>
     <div class="section-experiences">
       <div class="container-page">
-        <div class="section-experiences__content">
+        <header class="section-experiences__title">
+          <BaseHeading primary :label="$t('experiences.subtitle')">{{ $t('experiences.title') }}</BaseHeading>
+        </header>
+      </div>
+      <div class="section-experiences__content">
+        <div class="container-page">
           <StoryblokComponent v-for="blok in bloks" :key="blok._uid" :blok="blok" class="section-experiences__blok" />
           <EndOfPage />
         </div>
@@ -18,6 +18,7 @@
 </template>
 
 <script setup lang="ts">
+import type { SbBlokData } from '@storyblok/js'
 import { useLocalizedStoryParams } from '@/composables/useLocalizedStoryParams'
 
 defineOptions({
@@ -36,8 +37,8 @@ await useAsyncData('jobs', () => store.fetchJobs())
 
 const content = computed(() => page.value.content)
 const bloks = computed(() =>
-  content.value.body.map((item) => {
-    if (item.component === 'ItemsList') item.items = store.jobs
+  content.value.body.map((item: SbBlokData) => {
+    if (item.component === 'ItemsList' && store.jobs?.length) item.items = store.jobs
     if (item.component === 'LongText') item.component = 'ExperienceSummary'
 
     return item
@@ -49,8 +50,8 @@ $setMetadata($getMetadataFromStory(content.value))
 
 <style lang="scss">
 .section-experiences {
-  background-color: var(--color-main-lightest);
   &__content {
+    background-color: var(--color-main-lightest);
     display: grid;
     .section-title {
       margin-bottom: 2rem;