+++ pages/publications/index.vue
@@ -2,10 +2,11 @@
   <div class="section-publications">
     <div class="container-page">
       <header class="section-publications__title">
-        <BaseHeading primary :label="$t('publications.subtitle')">{{ $t('publications.title') }}</BaseHeading>
+        <BaseHeading primary :label="$t('publications.subtitle')">{{ content.title }}</BaseHeading>
+        <BlogSections v-if="blogSections.length" :pages="blogSections" />
       </header>
       <div class="section-publications__content">
-        <PostsList :posts="posts" />
+        <PostsList v-if="postsWithContent.length" :posts="postsWithContent" />
         <EndOfPage />
       </div>
     </div>
@@ -21,6 +22,10 @@ defineOptions({
   name: 'PublicationsPage',
 })
 
+defineI18nRoute({
+  locales: ['it', 'en'],
+})
+
 const storyblokApi = useStoryblokApi()
 const { $getMetadataFromStory, $setMetadata } = useNuxtApp()
 
@@ -28,16 +33,31 @@ const { getParams } = useLocalizedStoryParams()
 
 const page = await useAsyncStoryblok('publications', getParams()).catch(() => null)
 
+if (!page) {
+  throw createError({
+    statusCode: 404,
+  })
+}
+
 const posts = await storyblokApi
   .get(`cdn/stories`, {
-    ...getParams({ withFallback: true }),
-    is_startpage: false,
+    ...getParams(),
     starts_with: 'publications/',
+    level: 2,
     sort_by: 'first_published_at:desc',
   })
   .then((response) => response.data.stories ?? [])
+  .catch(() => null)
 
 const content = computed(() => page.value.content)
+
+const postsWithContent = computed(() => {
+  return posts.filter((e) => !e.is_startpage && e.content.title)
+})
+
+const blogSections = computed(() => {
+  return posts.filter((e) => e.is_startpage && e.content.title)
+})
 
 $setMetadata($getMetadataFromStory(content.value))
 </script>
@@ -50,6 +70,7 @@ $setMetadata($getMetadataFromStory(content.value))
       grid-template-areas: 'title content';
       grid-template-columns: 40% auto;
       height: 100vh;
+      gap: 3rem;
     }
     &__content {
       grid-area: content;